requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0
pydantic>=2.5.0
pyyaml>=6.0
psutil>=5.9.0
python-multipart>=0.0.6
rich>=13.0.0